+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ### 0.0.3
 
-stringId base can be customized in an input box
+* stringId base can be customized in an input box
+* New function: **Get String Contents**
+  Gathers all stringIds and correcponding texts from a mission file and copies them for further usage in the CNC translations table
 
 ### 0.0.2